requirements_api.txt
aiohttp==3.8.4
aiohttp-swagger3==0.7.4
astropy==5.2.1
bcrypt==4.0.1
cryptography==39.0.1
gunicorn==20.1.0
matplotlib==3.7.0
motor==3.1.1
multidict==6.0.4
numba==0.56.4
numpy==1.23.5
odmantic==0.9.2
pandas==1.5.3
PyJWT==2.6.0
pymongo==4.3.3
pytest==7.2.1
pytest-aiohttp==1.0.4
PyYAML==6.0
requests==2.28.2
sshtunnel==0.4.0
supervisor==4.2.5
uvloop==0.17.0
